requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
boto3==1.35.36
botocore==1.35.36
groq
python-multipart==0.0.12
httpx